Add in similarity theory flux computation
